fix: refactor TestEntityCreator to be in Entities package
also added optional parameters to getTestCar
