+++ src/test/java/entities/TestEntityCreator.java
@@ -1,4 +1,4 @@
-package entitypackagers;
+package entities;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
@@ -8,7 +8,7 @@ import attributes.AttributeMap;
 import constants.EntityStringNames;
 import constants.Exceptions;
 
-import entities.*;
+import entitypackagers.AttributizeLoanDataUseCase;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -17,16 +17,20 @@ import java.util.Map;
 
 public class TestEntityCreator {
 
-    static Car getTestCar() {
+    public static Car getTestCar() {
+        return getTestCar(50000.0, "Honda", "Civic", 2020.0, 100.0, new AttributeMap(), 3.0);
+    }
+
+    public static Car getTestCar(double price, String make, String model, double year, double kms, AttributeMap addOns, double id) {
         Car car = null;
         AttributeMap carMap = new AttributeMap();
-        carMap.addItem(EntityStringNames.CAR_PRICE, 50000.0);
-        carMap.addItem(EntityStringNames.CAR_MAKE, "Honda");
-        carMap.addItem(EntityStringNames.CAR_MODEL, "Civic");
-        carMap.addItem(EntityStringNames.CAR_YEAR, 2020.0);
-        carMap.addItem(EntityStringNames.CAR_KILOMETRES, 100.0);
-        carMap.addItem(EntityStringNames.ADD_ON_STRING, new AttributeMap());
-        carMap.addItem(EntityStringNames.CAR_ID, 3.0);
+        carMap.addItem(EntityStringNames.CAR_PRICE, price);
+        carMap.addItem(EntityStringNames.CAR_MAKE, make);
+        carMap.addItem(EntityStringNames.CAR_MODEL, model);
+        carMap.addItem(EntityStringNames.CAR_YEAR, year);
+        carMap.addItem(EntityStringNames.CAR_KILOMETRES, kms);
+        carMap.addItem(EntityStringNames.ADD_ON_STRING, addOns);
+        carMap.addItem(EntityStringNames.CAR_ID, id);
         try {
             car = CarFactory.getEntity(carMap);
         } catch (Exceptions.FactoryException e) {
@@ -35,7 +39,7 @@ public class TestEntityCreator {
         return car;
     }
 
-    static CarBuyer getTestBuyer() {
+    public static CarBuyer getTestBuyer() {
         CarBuyer buyer = null;
         AttributeMap buyerMap = new AttributeMap();
         buyerMap.addItem(EntityStringNames.BUYER_BUDGET, 30000.5);
@@ -49,7 +53,7 @@ public class TestEntityCreator {
         return buyer;
     }
 
-    static AddOn getTestAddOn() {
+    public static AddOn getTestAddOn() {
         AddOn addOn = null;
         AttributeMap addOnMap = new AttributeMap();
         addOnMap.addItem(EntityStringNames.ADD_ON_NAME, "Rust proofing");
@@ -63,7 +67,7 @@ public class TestEntityCreator {
         return addOn;
     }
 
-    static LoanData getTestLoanData() {
+    public static LoanData getTestLoanData() {
         LoanData loan = null;
         AttributeMap loanMap = new AttributeMap();
         loanMap.addItem(EntityStringNames.LOAN_AMOUNT, 1.25);
