+++ src/test/java/entitypackagers/AttributizeCarBuyerUseCaseTest.java
@@ -8,6 +8,7 @@ import constants.EntityStringNames;
 
 import entities.CarBuyer;
 
+import entities.TestEntityCreator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
