+++ src/test/java/entitypackagers/AttributizeCarUseCaseTest.java
@@ -9,6 +9,7 @@ import constants.EntityStringNames;
 import entities.AddOn;
 import entities.Car;
 
+import entities.TestEntityCreator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
