+++ src/test/java/entitypackagers/AttributizeLoanDataUseCaseTest.java
@@ -9,6 +9,7 @@ import constants.EntityStringNames;
 
 import entities.LoanData;
 
+import entities.TestEntityCreator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
